[DFA] Add a proper determinisation algorithm
The first implementation suffered from a flaw where, if one rule
was a subset of another and it contained a repetition node, it would
become stuck on an infinite loop because it would try to unroll
the node by creating two nodes that would be equivalent to the
one being determinised.

The new implementation instead properly keep track of the node
combinations that have been handled, and map to existing node even
if the route to them is different from the first time they were
encountered.