[BOOTSTRAP] Add partial support for strings
In order to support spaces in patters, we must first have string
litterals supports. It is, in the current version, quite limited in
that it does not include escapes, including quotes ones, meaning only
one litteral could be used per line. Proper support for this feature
will require [^"] to be working.